Switch to leader f for opening fzf.
I was finding that having files under ctrl-p and buffers under leader b
was causing me to use the file finder in places where the buffer would
be better.

Trying to put them both under to leader to see if that works.

I moved ALE fix to leader x, but I don't use it that often.