Set up omnicomplete using language servers for completion.
I want to be able to auto complete things like people with fancy newer
IDEs. Language servers allow for things like going to definitions n'
stuff, but it also has the ability for autocomplete.

Pretty fun stuff. I managed to get it set up for rust, typescript and
python. The TypeScript language server appears to be able to manage
JavaScript as well.