Add a merge PR alias.
This is lifted from Felix Crux. It beats the GitHub "rebase and merge"
because I can keep my commit signing. The downside is that it pushes to
master which could be locked for some repos. I would like it if there
was a confirm step, but that can be added later.

I also move my `edit-conflicts` alias into a git alias.