Update the way that NVM is setup.
I was having an annoying issue where when creating a new shell instance
a `./~` directory would be created. I tracked it down to nvm by commenting things
out and adding them back slowing and seeing if the `~` directories were
still being created.

This alternative way seems to be the current suggested approach in the
nvm docs but I don't really know "why" this solves the issue.