Add an alias to modernize branch names.
I feel that `master` is not an acceptable term. This is supported by
IETF  https://tools.ietf.org/id/draft-knodel-terminology-00.html#master-slave

This alias will update the branch and will leave the old branch in
place. The manual process of updating the default branches in GitHub
still need to be done.