Change installation method away from symlinks
I have been having issues with symlinks and vim. One issue is that
weird locally directories are created call `~` when vim is started. The
other issue is that `Coc` plugins don't work as expected.

This commit switches to useing a new install method using `rsync` rather
than symlinks. Some things in the XDG_CONFIG directory are still using
symlinks, but I may need to switch that.