Add zoxide a directory jumper.
I used the shell script z for quite some time. I ended up removing it
when I moved to the "non-symlinking" approach to bootstrapping the repo.
I found that I was missing the command but would like to use some of the
newer tools that people are building in Rust.